Création de l'ID3D12Device
    Inclure les entêtes nécessaires:
        d3d12.h: pour les fonctions et interfaces Direct3D 12.
        dxgi1_6.h: pour les versions plus récentes de DXGI (interface pour la gestion des périphériques graphiques et des écrans).
        wrl.h: pour utiliser ComPtr, un pointeur intelligent pour gérer les interfaces COM.

    Définir un pointeur pour ID3D12Device:
    Utilisez Microsoft::WRL::ComPtr<ID3D12Device> mD3DDevice; pour définir un pointeur intelligent qui gère automatiquement les références d'objet COM.

    Créer le périphérique Direct3D:
    Utilisez la fonction D3D12CreateDevice pour initialiser le périphérique. Cette fonction prend en entrée l'adaptateur (dans notre cas, l'adaptateur par défaut avec nullptr), le niveau de fonctionnalités souhaité (nous avons utilisé D3D_FEATURE_LEVEL_11_0), et un pointeur à l'interface du périphérique créé.

    Vérification:
    Vérifiez que la création du périphérique s'est déroulée avec succès en examinant la valeur de retour de D3D12CreateDevice. De plus, assurez-vous que votre GPU prend en charge Direct3D 12.
