+++ Engine/main.h
@@ -1,0 +1,3 @@
+#pragma once
+
+#include "resource.h"
