+++ Engine/x64/Debug/Engine.log
@@ -1,2 +1,2 @@
 ﻿  Engine.cpp
-  Engine.vcxproj -> C:\Users\mhirth\source\repos\GameEngine\x64\Debug\Engine.exe
+  Engine.vcxproj -> C:\Users\lprieu\Documents\repo\GTech3-DirectX-Engine\x64\Debug\Engine.exe
