+++ Engine/x64/Debug/Engine.tlog/Cl.items.tlog
@@ -1,1 +1,2 @@
-C:\Users\mhirth\source\repos\GameEngine\Engine\Engine.cpp;C:\Users\mhirth\source\repos\GameEngine\Engine\x64\Debug\Engine.obj
+C:\Users\lprieu\Documents\repo\GTech3-DirectX-Engine\Engine\Engine.cpp;C:\Users\lprieu\Documents\repo\GTech3-DirectX-Engine\Engine\x64\Debug\Engine.obj
+C:\Users\lprieu\Documents\repo\GTech3-DirectX-Engine\Engine\main.cpp;C:\Users\lprieu\Documents\repo\GTech3-DirectX-Engine\Engine\x64\Debug\main.obj
