+++ Game/x64/Debug/Game.exe.recipe
@@ -2,7 +2,7 @@
 <Project>
   <ProjectOutputs>
     <ProjectOutput>
-      <FullPath>C:\Users\mhirth\source\repos\GameEngine\x64\Debug\Game.exe</FullPath>
+      <FullPath>C:\Users\lprieu\Documents\repo\GTech3-DirectX-Engine\x64\Debug\Game.exe</FullPath>
     </ProjectOutput>
   </ProjectOutputs>
   <ContentFiles />
