+++ Game/x64/Debug/Game.log
@@ -1,2 +1,2 @@
 ﻿  Game.cpp
-  Game.vcxproj -> C:\Users\mhirth\source\repos\GameEngine\x64\Debug\Game.exe
+  Game.vcxproj -> C:\Users\lprieu\Documents\repo\GTech3-DirectX-Engine\x64\Debug\Game.exe
