+++ Game/x64/Debug/Game.tlog/Cl.items.tlog
@@ -1,1 +1,1 @@
-C:\Users\mhirth\source\repos\GameEngine\Game\Game.cpp;C:\Users\mhirth\source\repos\GameEngine\Game\x64\Debug\Game.obj
+C:\Users\lprieu\Documents\repo\GTech3-DirectX-Engine\Game\Game.cpp;C:\Users\lprieu\Documents\repo\GTech3-DirectX-Engine\Game\x64\Debug\Game.obj
