+++ Game/x64/Debug/Game.tlog/Game.lastbuildstate
@@ -1,2 +1,2 @@
 PlatformToolSet=v143:VCToolArchitecture=Native64Bit:VCToolsVersion=14.37.32822:TargetPlatformVersion=10.0.22621.0:
-Debug|x64|C:\Users\mhirth\source\repos\GameEngine\|
+Debug|x64|C:\Users\lprieu\Documents\repo\GTech3-DirectX-Engine\|
